Refactor: extract tags into subdirectory & smaller files
