+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.8.121+1207
+version: 0.8.121+1208
 
 msix_config:
   display_name: Lotti
